apps/vts/requirements.txt
python-dotenv==1.0.1
websockets==13.1
asyncio==3.4.3
requests==2.32.3
pyyaml==6.0.2